Rename class into SafeParser
