+++ Manifest.txt
@@ -2,5 +2,5 @@ History.txt
 Manifest.txt
 README.txt
 Rakefile
-lib/hash_parser.rb
-test/test_hash_parser.rb
+lib/safe_parser.rb
+test/test_safe_parser.rb
